BMA-5963 Fix Docker usage on Linux
On Linux, the files from the repository are owned by 1000 (while they
are mapped to root on MacOS). This causes an issue when running git
commands from Docker as the users differ (root vs 1000) which causes git
to error out (since CVE fix in version v2.35.2).
The git command "git status --porcelain" is run by the go build to stamp
it with the VCS information. This is the default since Go 1.18. It can
be disabled with "-buildvcs=false". However there are still issue then
during the linting so the best fix is to mark the repository as safe